posts + comments scripts
This an update to the comments script that grabs all the comments from a submission(post) and turns it into a csv. I also added a posts script that grabs a number of posts from user input and puts it in a csv
