+++ posts.csv
@@ -1,0 +1,19 @@
+submission(post),upvotes,downvotes
+"My older brother is so annoying and he does so much stuff that annoys me that i would need a few hours to explain it all so im just gonna briefly go over it in a list
+-Threatened me with an airsoft gun and pointed it at the neighbors and nothing happened to him
+-He takes so much food from the kitchen theres never anything nice that lasts longer then 2 seconds
+-He once sprained my ankle while ""playfighting"" with me and my mum couldnt have gave less of a shit about punishing him.
+-He feels entitled to take and break my stuff
+-He hits me whenever and wherever but he hasnt done it recently
+-He often does obnoxious things like throwing those crackerbombs (i think thats what its called) at me while im reading even though it says on the packet its dangerous to throw at people
+-He calls me a gay retard and constantly uses the n word
+-He litters and doesnt care
+-I was diagnosed with aspergers and my mum told my brother about it and when i got mad at her for telling him without fucking asking me she said it will help him be more supportive and treat me differently.She wasnt wrong about treating me differently because now his favourite insult to use on me is Disabled retard.
+-We go to the same school so now everyone assosiates me with him and thats pretty much all im known for
+-While we are on holiday his favouritr thing to do is put me on an infatable water mattress and push me out to sea because he knows i cant swim and i cant get back without him so he says that he will only bring me back if i say dumb stuff like ""im a gay retard""
+-He gets everything better then me
+-He loves to sit on me which hurts my stomach so much
+
+When i ask my mum to have a word with him or punish him she tells me its all normal because hes a teenager and im being overly dramatic.I cant speak to my dad either because my parents are divorced and my dad moved out.Im fed up of coming home feeling like shit and suffering from his repeated insults.I dont have anyone to talk to ever because nobody would care.",39.0,0.0
+"I was playing Minecraft mobile 5 years ago, and I hadn’t named my worlds. I was checking which world was which, so I could delete the correct one. What I didn’t know is that worlds are sorted by recent. I deleted my favorite world. The tears on my face...",32.0,0.0
+"A couple of days ago I went to drop of my fiancé to work (8hours away) I ended up staying there for two nights at a hotel. The first night was good and then the second day when my fiancé was at work and it was time for me to pack up to go back home. I started packing up and then I realized I couldn’t find my iPhone charger or my Apple Watch. I went outside to look in my car and locked myself out of the room. I went to the front desk where the person who was there was not the most welcoming person, he seemed like I was bothering just for trying to get a key for my room. Then he sent another employee to open the door for me instead of giving me a key I thought it was weird but didn’t want to think much of it. Then he proceeded to ask if my fiancée was with me which I didn’t like. He then called me (mami) while opening the door for me and asking if I was going to be there at night.... I freaked out said no and locked myself inside. 2 minutes later he comes knocking at my door (i do not open) and asks if the car outside was mine. It might not seem like a big deal to some people but I freaked out and my heart started racing. I did not feel comfortable at all and wanted to leave because I was scared and I wasn’t able to reach my fiancé because of bad signal. Before getting my things I looked outside the window to make sure the guy was gone and got my things and got in my car. I forgot anything about my charger or my Apple Watch and I left. I barely had any money and I had to buy an expensive charger since my phone was almost dead and I was 8 hours away from home. I’m not the type of person that has a lot of money so it really sucked that my Apple Watch was nowhere to be found and it’s offline and the last location on it was the hotel...",18.0,0.0
